+++ _Typings/editorApi/modules/console.d.ts
@@ -4,21 +4,21 @@
 declare module "editor:module" {
     namespace console {
         /**
-         * Log a message to the console.
+         * Logs a message to the console.
          * @param data - The data to log.
          * @returns
          */
         const log: (...data: any[]) => void
 
         /**
-         * Log an error message to the console.
+         * Logs an error message to the console.
          * @param data - The data to log.
          * @returns
          */
         const error: (...data: any[]) => void
 
         /**
-         * Clear the console.
+         * Clears the console.
          * @returns
          */
         const clear: () => void