+++ _Typings/utils/basicData.d.ts
@@ -15,6 +15,11 @@ declare module "basicData" {
 
             x: float
             y: float
+
+            /**
+             * @param x - The x component of the float2.
+             * @param y - The y component of the float2.
+             */
             constructor(x: float, y: float)
         }
 
@@ -24,6 +29,12 @@ declare module "basicData" {
             x: float
             y: float
             z: float
+
+            /**
+             * @param x - The x component of the float3.
+             * @param y - The y component of the float3.
+             * @param z - The z component of the float3.
+             */
             constructor(x: float, y: float, z: float)
         }
 
@@ -34,6 +45,13 @@ declare module "basicData" {
             y: float
             z: float
             w: float
+
+            /**
+             * @param x - The x component of the float4.
+             * @param y - The y component of the float4.
+             * @param z - The z component of the float4.
+             * @param w - The w component of the float4.
+             */
             constructor(x: float, y: float, z: float, w: float)
         }
 
@@ -44,6 +62,13 @@ declare module "basicData" {
             y: float
             z: float
             w: float
+
+            /**
+             * @param x - The x component of the quaternion.
+             * @param y - The y component of the quaternion.
+             * @param z - The z component of the quaternion.
+             * @param w - The w component of the quaternion.
+             */
             constructor(x: float, y: float, z: float, w: float)
         }
 
@@ -53,6 +78,12 @@ declare module "basicData" {
             x: bool
             y: bool
             z: bool
+
+            /**
+             * @param x - The x component of the bool3.
+             * @param y - The y component of the bool3.
+             * @param z - The z component of the bool3.
+             */
             constructor(x: bool, y: bool, z: bool)
         }
 
@@ -62,6 +93,12 @@ declare module "basicData" {
             pos: Float3
             rot: Float3
             scl: Float3
+
+            /**
+             * @param pos - The position component of the transform.
+             * @param rot - The rotation component of the transform.
+             * @param scl - The scale component of the transform.
+             */
             constructor(pos: Float3, rot: Float3, scl: Float3)
         }